+++ backend/service-lib-server/src/main/kotlin/micro/Module.kt
@@ -1,7 +1,6 @@
 package dk.sdu.cloud.micro
 
 import dk.sdu.cloud.ServiceDescription
-import dk.sdu.cloud.debug.DebugSystem
 import dk.sdu.cloud.debug.DebugSystemFeature
 import dk.sdu.cloud.faults.FaultInjectionFeature
 import dk.sdu.cloud.service.CommonServer
@@ -88,6 +87,7 @@ class ServiceRegistry(
                 install(BackgroundScopeFeature)
                 install(DebugSystemFeature)
                 install(ScriptManager)
+                install(PrometheusFeature)
             }
         }
     }